BREAKING: initial support for links, changes to denote-db-query
Links to other notes will now be cached.  They are stored as IDs
separated by whitespaces in the database.  You should delete your
previous database and recreate it afterwards.

I used this chance to introduce a breaking change in the API:
keywords (and now links) are splitted at whitespaces and returned as
lists.  This way there's no need to do that manually, which
complicated real-world use cases including the examples in the README.
Also, the new :no-dup key allow you to delete duplicate entries.